Pulled the text twiddling out into a separate file
