+++ src/extension.ts
@@ -10,6 +10,7 @@ import * as batch from './batch';
 import * as shell from './shell';
 import * as azurebatchtree from './azurebatchtree';
 import * as host from './host';
+import * as textmodels from './textmodels';
 
 let diagnostics : vscode.DiagnosticCollection;
 
@@ -345,9 +346,6 @@ export async function convertToParameterCore(document: vscode.TextDocument, sele
 
     const propertyType = getParameterTypeName(propertyValue); // consider getting this from Swagger?
 
-    const parametersElement = jsonSymbols.find((s) => s.name == 'parameters' && !s.containerName);
-    const needToCreateParametersElement = !parametersElement;
-
     // TODO: investigate using a smart insert for this (https://github.com/Microsoft/vscode/issues/3210)
     // (Currently doesn't seem to be a thing - works for completion items only...?)
     const newParameterDefn : any = {
@@ -356,133 +354,19 @@ export async function convertToParameterCore(document: vscode.TextDocument, sele
         metadata: { description: `Value for ${property.containerName}.${property.name}` }
     }
 
-    const model = createParameterConversionModel(jsonSymbols, property.name, newParameterDefn);
-
-    const insertParamDefn : vscode.TextEdit = model.getEdit();
+    const insertParamEdit = textmodels.getTemplateParameterInsertion(jsonSymbols, property.name, newParameterDefn);
 
     const replaceValueWithParamRef = new vscode.TextEdit(propertyValueLocation, ` "[parameters('${property.name}')]"`);
 
+    await applyEdits(document, insertParamEdit, replaceValueWithParamRef);
+
+    return insertParamEdit;
+}
+
+async function applyEdits(document : vscode.TextDocument, ...edits : vscode.TextEdit[]) : Promise<boolean> {
     const wsEdit = new vscode.WorkspaceEdit();
-
-    wsEdit.set(document.uri, [insertParamDefn, replaceValueWithParamRef]);
-    await vscode.workspace.applyEdit(wsEdit);
-
-    return insertParamDefn;
-}
-
-interface IParameterConversionModel {
-    getEdit() : vscode.TextEdit;
-}
-
-function indentOf(symbol : vscode.SymbolInformation) {
-    return symbol.location.range.start.character;
-}
-
-function indentFrom(symbol1 : vscode.SymbolInformation, symbol2 : vscode.SymbolInformation) {
-    const offset = indentOf(symbol1) - indentOf(symbol2);
-    return Math.abs(offset);
-}
-
-function indentLines(text : string, amount : number) : string {
-    const indent = ' '.repeat(amount);
-    const lines = text.split('\n');
-    const indented = lines.map((l) => indent + l);
-    return indented.join('\n');
-}
-
-function isSingleLine(range: vscode.Range) {
-    return range.start.line == range.end.line;
-}
-
-function lineStart(pos: vscode.Position) {
-    return new vscode.Position(pos.line, 0);
-}
-
-function immediatelyBefore(pos: vscode.Position) {
-    return pos.translate(0, -1);
-}
-
-function createParameterConversionModel(jsonSymbols : vscode.SymbolInformation[], propertyName : string, newParameterDefn : any) : IParameterConversionModel {
-    const parametersElement = jsonSymbols.find((s) => s.name == 'parameters' && !s.containerName);
-    const parameterSymbols = jsonSymbols.filter((s) => s.containerName == 'parameters');
-    const lastExistingParameter = parameterSymbols.length > 0 ? parameterSymbols.reverse()[0] : undefined;  // not sure what order guarantees the symbol provider makes, but it's not critical if this isn't actually the last one
-
-    class InsertAfterExistingParameter implements IParameterConversionModel {
-        constructor(private readonly parametersElement : vscode.SymbolInformation, private readonly existingParameter : vscode.SymbolInformation) {
-        }
-        getEdit() : vscode.TextEdit {
-            const indentPerLevel = indentFrom(this.existingParameter, this.parametersElement);
-            const initialIndent = indentPerLevel * 2;
-            const rawNewParameterDefnText = `"${propertyName}": ${JSON.stringify(newParameterDefn, null, indentPerLevel)}`;
-            const newParameterDefnText = indentLines(rawNewParameterDefnText, initialIndent);
-            const insertText = ',\n' + newParameterDefnText;
-
-            const insertPos = this.existingParameter.location.range.end;
-            const edit = vscode.TextEdit.insert(insertPos, insertText);
-
-            return edit;
-        }
-    }
-
-    class InsertIntoEmptyParametersSection implements IParameterConversionModel {
-        constructor(private readonly parametersElement : vscode.SymbolInformation) {
-        }
-        getEdit() : vscode.TextEdit {
-            const indentPerLevel = this.parametersElement.location.range.start.character;
-            const initialIndent = indentPerLevel * 2;
-
-            const rawNewParameterDefnText = `"${propertyName}": ${JSON.stringify(newParameterDefn, null, indentPerLevel)}`;
-            const newParameterDefnText = indentLines(rawNewParameterDefnText, initialIndent);
-            const parametersElementOnOneLine = isSingleLine(this.parametersElement.location.range);  // for the "parameters": {} case
-
-            // prefix and suffix are for fixing up cases where the parameters element is squashed on one line
-            const prefix = (parametersElementOnOneLine ? '\n' : '');
-            const suffix = (parametersElementOnOneLine ? (' '.repeat(indentPerLevel)) : '');
-            const insertText = prefix + newParameterDefnText + '\n' + suffix;  // TODO: line ending
-
-            // if the parameters element is squashed then our insert position is to the left of the closing brace (which will push things into the right place)
-            // otherwise we want to insert at the start of the line containing the closing brace
-            const closingBracePos = this.parametersElement.location.range.end;
-            const insertPos = (parametersElementOnOneLine ? immediatelyBefore(closingBracePos) : lineStart(closingBracePos));
-            const edit = vscode.TextEdit.insert(insertPos, insertText);
-
-            return edit;
-        }
-    }
-
-    class InsertNewParametersSection implements IParameterConversionModel {
-        getEdit() : vscode.TextEdit {
-            let parameters : any = {};
-            parameters[propertyName] = newParameterDefn;
-
-            const topLevelElement = jsonSymbols.find((s) => !s.containerName);
-            const indentPerLevel = topLevelElement ? indentOf(topLevelElement) : 2;
-            const rawParametersSection = `"parameters": ${JSON.stringify(parameters, null, indentPerLevel)}`;
-
-            const parametersSectionText = indentLines(rawParametersSection, indentPerLevel);  // going in at top level so indent only once
-            const topLevelElementSeparator = topLevelElement ? ',\n' : '\n';
-            const insertText = parametersSectionText + topLevelElementSeparator;
-            
-            const insertPos = new vscode.Position(1, 0);
-            const edit = vscode.TextEdit.insert(insertPos, insertText);
-            
-            return edit;
-        }
-
-    }
-
-    if (parametersElement) {
-        if (lastExistingParameter) {
-            // We are appending after an existing parameter
-            return new InsertAfterExistingParameter(parametersElement, lastExistingParameter);
-        } else {
-            // We are inserting into an empty parameters section
-            return new InsertIntoEmptyParametersSection(parametersElement);
-        }
-    } else {
-        // There is no parameters section - we need to create one and insert it at the top of the document
-        return new InsertNewParametersSection();
-    }
+    wsEdit.set(document.uri, edits);
+    return await vscode.workspace.applyEdit(wsEdit);
 }
 
 async function getJsonSymbols(document : vscode.TextDocument) : Promise<vscode.SymbolInformation[]> {
